路由更新

+++ goperation/manager/wsgi/file/controller.py
@@ -52,24 +52,54 @@ FAULT_MAP = {InvalidArgument: webob.exc.HTTPClientError,
 class FileReuest(BaseContorller):
 
 
-    def index(self, req, body):
+    def index(self, req):
         session = get_session(readonly=True)
         query = model_query(session, DownFile)
         return resultutils.results(result='list file success',
                                    data=[downfile.to_dict() for downfile in query.all()])
 
-    def show(self, req, file_id, body):
+    def create(self, req, body):
+        try:
+            address = body.pop('address')
+            size = body.pop('size')
+            md5 = body.pop('md5')
+            crc32 = body.pop('crc32')
+        except KeyError:
+            raise InvalidArgument('Add file miss some keywords')
+        ext = body.get('ext')
+        if not ext:
+            ext = address.split('.')[-1]
+        downfile = DownFile(md5=md5,
+                            crc32=crc32,
+                            downloader=body.get('downloader', 'http'),
+                            adapter_args=body.get('adapter_args'),
+                            address=address,
+                            ext=ext,
+                            size=size,
+                            desc=body.get('desc'),
+                            uploadtime=body.get('uploadtime', timeutils.utcnow())
+                            )
+        return resultutils.results(result='Add file success', data=[dict(uuid=downfile.uuid,
+                                                                         md5=downfile.md5,
+                                                                         crc32=downfile.crc32,
+                                                                         size=downfile.size,
+                                                                         uploadtime=downfile.uploadtime,
+                                                                         downloader=downfile.downloader)])
+
+    def show(self, req, file_id):
         session = get_session(readonly=True)
         query = model_query(session, DownFile)
         if uuidutils.is_uuid_like(file_id):
-            query = query.filter(uuid=file_id)
+            query = query.filter_by(uuid=file_id)
         elif jsonutils.is_md5_like(file_id):
-            query = query.filter(md5=file_id)
+            query = query.filter_by(md5=file_id)
         elif jsonutils.is_crc32_like(file_id):
-            query = query.filter(crc32=file_id)
+            query = query.filter_by(crc32=file_id)
         else:
             raise InvalidArgument('File id not uuid or md5 or crc32')
-        downfile = query.one()
+        downfile = query.one_or_none()
+        if not downfile:
+            return resultutils.results(resultcode=manager_common.RESULT_ERROR, result='Get file fail, no found')
         file_info = {'downloader': downfile.downloader,
                      'address': downfile.address,
                      'ext': downfile.ext,
@@ -82,45 +112,31 @@ class FileReuest(BaseContorller):
             file_info.setdefault('adapter_args', downfile.adapter_args)
         if downfile.desc:
             file_info.setdefault('desc', downfile.desc)
-        return file_info
-
-    def create(self, req, body):
-        address = body.get('address')
-        if not address:
-            raise InvalidArgument('Add file miss address')
-        ext = body.get('ext')
-        if not ext:
-            ext = address.split('.')[-1]
-        downfile = DownFile(md5=body.get('md5'),
-                            crc32=body.get('crc32'),
-                            downloader=body.get('downloader', 'http'),
-                            adapter_args=body.get('adapter_args'),
-                            address=address,
-                            ext=ext,
-                            size=body.get('size'),
-                            desc=body.get('desc'),
-                            uploadtime=body.get('uploadtime', timeutils.utcnow())
-                            )
-        return resultutils.results(result='Add file success', data=[downfile.to_dict()])
+        return resultutils.results(result='Get file success', data=[downfile, ])
 
     def delete(self, req, file_id, body):
         session = get_session()
         query = model_query(session, DownFile)
         if uuidutils.is_uuid_like(file_id):
-            query = query.filter(uuid=file_id)
+            query = query.filter_by(uuid=file_id)
         elif jsonutils.is_md5_like(file_id):
-            query = query.filter(md5=file_id)
+            query = query.filter_by(md5=file_id)
         elif jsonutils.is_crc32_like(file_id):
-            query = query.filter(crc32=file_id)
+            query = query.filter_by(crc32=file_id)
         else:
             raise InvalidArgument('File id not uuid or md5 or crc32')
-        with session.begin():
-            downfile = query.one()
-            count = query.delete()
-        if count:
-            return resultutils.results(result='Delete file success', data=[downfile.to_dict()])
-        return resultutils.results(resultcode=manager_common.RESULT_ERROR,
-                                   result='Delete file success')
+        with session.begin(subtransactions=True):
+            downfile = query.one_or_none()
+            if not downfile:
+                return resultutils.results(result='Delete file do nothing, not found')
+            query.delete()
+
+        return resultutils.results(result='Delete file success', data=[dict(uuid=downfile.uuid,
+                                                                            md5=downfile.md5,
+                                                                            crc32=downfile.crc32,
+                                                                            size=downfile.size,
+                                                                            uploadtime=downfile.uploadtime,
+                                                                            downloader=downfile.downloader)])
 
     @BaseContorller.AgentsIdformater
     def send(self, req, agent_id, file_id, body):
@@ -132,13 +148,13 @@ class FileReuest(BaseContorller):
     @BaseContorller.AgentIdformater
     def list(self, req, agent_id, body):
         """call by client, and asyncrequest
-        send file to agents
+        list file in agents
         """
         raise NotImplementedError
 
     @BaseContorller.AgentIdformater
     def clean(self, req, agent_id, file_id, body):
         """call by client, and asyncrequest
-        send file to agents
+        delete file from agents
         """
         raise NotImplementedError
